base1/iaa_pipe_filter/to_int_filter.go
package iaa_pipe_filter

import (
	"errors"
	"strconv"
)

/**
 * @author  wjj
 * @date  2020/9/8 1:40 上午
 * @description
 */
var ToIntFilterWrongFormatError = errors.New("input data should be []string")

type ToIntFilter struct {
}

func NewToIntFilter() *ToIntFilter {
	return &ToIntFilter{}
}
func (sf *ToIntFilter) Process(data Request) (Response, error) {
	parts, ok := data.([]string)
	if !ok {
		return nil, ToIntFilterWrongFormatError
	}
	// ret := []int{}
	var ret []int
	for _, part := range parts {
		s, err := strconv.Atoi(part)
		if err != nil {
			return nil, err
		}
		ret = append(ret, s)
	}
	return ret, nil
}
